Refactoring BaseSparseEncodingIT and enabling BWC tests
